refacto: Split main code into lib and main for cli
